Icon implemention done
