+++ server/routes/apiMember.ts
@@ -19,10 +19,10 @@ apiMemberRouter.use(express.json());
 
 apiMemberRouter.get("/", async (req: express.Request, res: express.Response) => {
     try {
-        let data : Member[] = await prisma.member.findMany();
+        let data: Member[] = await prisma.member.findMany();
         res.json(data);
     } catch (e: any) {
-        res.json({error: e});
+        res.json({ error: e });
         console.log(e);
     }
 });
@@ -39,28 +39,28 @@ apiMemberRouter.post("/", async (req: express.Request, res: express.Response) =>
             }
         });
         res.json(await prisma.member.findMany());
-    } catch (e : any) {
-        res.json({error: e});
+    } catch (e: any) {
+        res.json({ error: e });
         console.log(e)
     }
 });
 
 apiMemberRouter.put("/:id", async (req: express.Request, res: express.Response) => {
     try {
-        let member : any = await prisma.member.findUnique({
+        let member: any = await prisma.member.findUnique({
             where: {
                 id: Number(req.params.id)
             }
         });
 
-        let updatedMember : Member = {
+        let updatedMember: Member = {
             id: member.id,
             firstname: req.body.firstname ? req.body.firstname : member.firstname,
             lastname: req.body.lastname ? req.body.lastname : member.lastname,
             email: req.body.email ? req.body.email : member.email,
             phone: req.body.phone ? req.body.phone : member.phone,
             signedUpDate: member.signedUpDate,
-            active: req.body.active ? req.body.active : member.active
+            active: req.body.active
         };
 
         await prisma.member.update({
@@ -70,8 +70,8 @@ apiMemberRouter.put("/:id", async (req: express.Request, res: express.Response) 
             data: updatedMember
         })
         res.json(await prisma.member.findMany());
-    } catch (e : any) {
-        res.json({error: e});
+    } catch (e: any) {
+        res.json({ error: e });
         console.log(e)
     }
 });
@@ -84,8 +84,8 @@ apiMemberRouter.delete("/:id", async (req: express.Request, res: express.Respons
             }
         });
         res.json(await prisma.member.findMany());
-    } catch (e : any) {
-        res.json({error: e});
+    } catch (e: any) {
+        res.json({ error: e });
         console.log(e)
     }
 });
